moved a bunch of stuff to a .env file
