+++ .gitignore
@@ -22,6 +22,8 @@ instance/
 *.pyc
 __pycache__/
 
+.env
+
 instance/
 .idea/
