+++ server/.env
@@ -1,3 +1,22 @@
 TEAM_NAME="Shums, Ryan, Evan"
 FLASK_ENV="development"
-FLASK_APP="mentions_crawler_flask"
+SQLALCHEMY_DATABASE_URI = "postgresql+psycopg2:///mentionscrawler"
+SQLALCHEMY_TRACK_MODIFICATIONS = False
+SECRET_KEY = "nOtSoSeCrEtKeY!!!"
+
+# Sendgrid stuff
+SENDGRID_API_KEY = "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"
+FROM_EMAIL = "ryannarine97@gmail.com"
+
+REDIS_URL = "redis://"
+
+SCHEDULER_KEY = "r%z2?os7z!IFN-T:GYT@vs^=t9SMSRfS{zH]9c^p@d]*^8G%v5{Cn0*1RtJvHz+"
+
+REDDIT_CLIENT_ID = "auo7pZGyIVaJhw"
+REDDIT_CLIENT_SECRET = "thAk1F93RSQC2uA_6d0xKYNntD8"
+REDDIT_USER_AGENT = "mentions_crawler:redditpart:madebyevanandryan"
+
+TWITTER_API_KEY = "cvk7jHmbc3Y08jLhrMnelgAeL"
+TWITTER_API_SECRET_KEY = "PQ4WXZmBBd6f9SyspCaCn1Q2xpdawTY2YnwqF7YfpdYni5jlg7"
+TWITTER_ACCESS_TOKEN = "125311068-VrV3rhIY01mWzFXLy6Xa0UvPFAhHZ4oC6rnf2Y5A"
+TWITTER_ACCESS_TOKEN_SECRET = "62lyn41C46GZ16RmHcEgI4f5VH3n4pdlQjKH790qh3khi"
