+++ server/config.py
@@ -1,5 +1,5 @@
 import os
-#TEAM_NAME = os.environ['TEAM_NAME']
-SQLALCHEMY_DATABASE_URI = "postgresql+psycopg2:///mentionscrawler"
-SQLALCHEMY_TRACK_MODIFICATIONS = False
-SECRET_KEY = "nOtSoSeCrEtKeY!!!"
+TEAM_NAME = os.environ["TEAM_NAME"]
+SQLALCHEMY_DATABASE_URI = os.environ["SQLALCHEMY_DATABASE_URI"]
+SQLALCHEMY_TRACK_MODIFICATIONS = os.environ["SQLALCHEMY_TRACK_MODIFICATIONS"]
+SECRET_KEY = os.environ["SECRET_KEY"]
