+++ server/mentions_crawler_apis/db.py
@@ -1,7 +1,10 @@
 from .celery import app
-from
+from ..constants import TOKEN_TAG
+from .constants import DB_CLEAN_URL
+from requests import post
 
-@app.task
+
+@app.task(name="db.clean")
 def clean_db(token: str):
     cookies = {TOKEN_TAG: token}
     clean_db_request = post(DB_CLEAN_URL, cookies=cookies)
